requirements.txt
# Core ML
ultralytics
paddlepaddle==2.6.2
# paddleocr==2.7.0.3  <-- Install manually with: pip install paddleocr==2.7.0.3 --no-deps

# Web & Server
fastapi
uvicorn
python-multipart
python-dotenv

# Data & DB
pymongo
dnspython
boto3
httpx
numpy<2.0.0

# OCR Logic & Utils (Required by PaddleOCR --no-deps)
opencv-python
shapely
pyclipper
lmdb
tqdm
imgaug
scikit-image
pillow
PyYAML
fpdf2
